Made populateEventInfo more reusable between fragments
