+++ app/src/main/java/com/capstone/event_finder/activities/MainActivity.java
@@ -16,8 +16,11 @@ import com.capstone.event_finder.fragments.ExploreFragment;
 import com.capstone.event_finder.fragments.FeedFragment;
 import com.capstone.event_finder.fragments.ProfileFragment;
 import com.capstone.event_finder.interfaces.FeedFragmentInterface;
+import com.capstone.event_finder.models.Event;
 import com.capstone.event_finder.network.EventViewModel;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonObject;
 import com.parse.ParseUser;
 
 public class MainActivity extends AppCompatActivity {
@@ -63,6 +66,50 @@ public class MainActivity extends AppCompatActivity {
         });
         bottomNavigationView.setSelectedItemId(R.id.action_feed);
     }
+
+    public void populateEventInfo(Event event, JsonObject temp) {
+        event.setName(temp.get("name").getAsString());
+        event.setDescription(temp.get("description").getAsString());
+        event.setImageUrl(temp.get("image_url").getAsString());
+        event.setTimeStart(temp.get("time_start").getAsString());
+        event.setId(temp.get("id").getAsString());
+        event.setEventSiteUrl(temp.get("event_site_url").getAsString());
+        event.setCategory(temp.get("category").getAsString());
+        checkAndSetEventEndTime(event, temp);
+        checkAndSetEventCost(event, temp);
+        formatAndSetEventLocation(temp, event);
+    }
+
+    private void checkAndSetEventCost(Event event, JsonObject temp) {
+        if (temp.get("cost").toString().matches("null")) {
+            event.setCost("N/A");
+        } else {
+            event.setCost("$" + temp.get("cost").getAsString() + "0");
+        }
+    }
+
+    private void checkAndSetEventEndTime(Event event, JsonObject temp) {
+        if (temp.get("time_end").toString().matches("null")) {
+            event.setTimeEnd(temp.get("time_start").getAsString());
+        } else {
+            event.setTimeEnd(temp.get("time_end").getAsString());
+        }
+    }
+
+    private void formatAndSetEventLocation(JsonObject temp, Event event) {
+        JsonArray formattedLocation = temp.getAsJsonObject("location").getAsJsonArray("display_address");
+        StringBuilder formattedLocationString = new StringBuilder();
+        for (int i = 0; i < formattedLocation.size(); i++) {
+            formattedLocationString.append(formattedLocation.get(i).getAsString()).append(" ");
+        }
+        event.setLocation(formattedLocationString.toString());
+    }
+
+
+
+
+
+
 
     private void initialCallToEventApi(FeedFragmentInterface feedFragment) {
         setListener(feedFragment);
